Added mod_pubsub_reaper.
This optional module cleans up the pubsub subscription and state
tables when subscriptions expire. See XEP-0060 section 12.18.

mod_pubsub and node_hometree both make calls into this module, whether
or not its loaded. When the module is not loaded cleanup is just not
done, but it will be restarted when the module is restarted.

This module has the following config options:

  * notify_time (default: 300) - Time in seconds before sending a
  pending expiry notification.

  * reap_orphans (default: false) - Whether or not to scan the
  pubsub_state and pubsub_subscription tables on startup looking for
  orphans. This is potentially expensive and shouldn't need to happen,
  but is available as an option just in case.